Refactor merge scheduling code to allow overrides (#114547)
This code refactors how the merge scheduler is configured to allow
different engine implementations to configure different merge schedulers.
