+++ server/src/main/java/org/elasticsearch/index/engine/ElasticsearchMergeScheduler.java
@@ -1,0 +1,27 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the "Elastic License
+ * 2.0", the "GNU Affero General Public License v3.0 only", and the "Server Side
+ * Public License v 1"; you may not use this file except in compliance with, at
+ * your election, the "Elastic License 2.0", the "GNU Affero General Public
+ * License v3.0 only", or the "Server Side Public License, v 1".
+ */
+
+package org.elasticsearch.index.engine;
+
+import org.apache.lucene.index.MergeScheduler;
+import org.elasticsearch.index.merge.MergeStats;
+import org.elasticsearch.index.merge.OnGoingMerge;
+
+import java.util.Set;
+
+public interface ElasticsearchMergeScheduler {
+
+    Set<OnGoingMerge> onGoingMerges();
+
+    MergeStats stats();
+
+    void refreshConfig();
+
+    MergeScheduler getMergeScheduler();
+}
